Fixed errors and preformance tunning.
I fixed an issue in the cross entropy algorithm when it gave a 0% probability for a given character, causing it to take a log2(0).
Now it adds a small fraction to prevent it from ever trying to take an invalid log.

I also normalized the object file to [0.0, 1.0] which fixed the problem of always choosing the same output.

I also adjusted the learning constant and some training variables to get better accuracy.